Clean up python bindings generator, add support for remaining data types.
This is a significant refactor of the python bindings generator, with
the aim of making the generated code a bit cleaner and of better hiding
many of our implementation details from the public API. To help prove
out the approach, it adds previously-missing support for data types like
sequences and maps.

We use the new `TypeUniverse` structure to iterate over all types used
in the interface, and generate various internal helper functions as
methods on our utility classes. This keeps them out of the public API
as seen by consumers.

For example, for each type that lowers into a `RustBuffer`, there is
a corresponding `RustBuffer.allocFrom{{ type_name }}` staticmethod for
lowering it and a `RustBuffer.consumeInto{{ type_name }}` for lifting it.